refactor: 상세페이지 타이틀 컴포넌트 폴더구조 및 파일명 수정
