+++ components/detail/title/index.tsx
