+++ pages/travel/[Id]/detail/index.tsx
@@ -3,13 +3,13 @@ import React from "react";
 
 import Article from "@/components/detail/article";
 import Header from "@/components/detail/header/index";
-import DetailTitle from "@/components/detail/title/DetailTitle";
+import Title from "@/components/detail/title";
 
 function Detail() {
   return (
     <div className="w-1200 flex flex-col bg-sky-50">
       <Header />
-      <DetailTitle />
+      <Title />
       <Article />
       <div
         className="fixed bottom-40 right-40 h-64 w-64 animate-bounce cursor-pointer"
